Rename: stores/selector/issueList -> issueListQuery
